extract swift extraction rules behind the indexer registry
Swift becomes the second language to own a module under
`indexer/src/languages/`: the rule file, the compiled-rule cache, the
member-call callsite marker, the receiver-call engine, and the projection
facts that used to be spread across `lib.rs`, `language_configs.rs` and
`semantic/mod.rs` now sit in one `LanguageExtraction` row. Every consumer
keeps reading "registry first, residual match second", so only Swift's arm
is deleted at each site and the registry and residual halves stay disjoint.

The bodies are moved verbatim apart from two renames that match the Kotlin
module's shape: `collect_swift_receiver_call_edges` -> `receiver_call_specs`
and `swift_member_call` -> `member_call`. 24 functions, 2 structs, 1 enum and
1 impl block left `lib.rs`; the same items landed in `languages/swift.rs`.

Two Swift facts are recorded as they are, not as they "should" be.
`route_comments_are_c_style` is `false` because the framework-route scanner
has never stripped `//` for Swift, while the CLI comment roster does dim
`//` for Swift and now reads it from the contracts comment profile. Those
are separate facts with separate owners; unifying them would change route
claims or rendering.

`tests/language_extraction_snapshot.rs` gains a Swift row and two golden
files, captured before the move and byte-identical after it. Mutation proof:
dropping `receiver_call_specs`, flipping
`promotes_type_member_functions_to_methods`, or changing
`qualified_name_delimiter` each fail the new snapshots, and the delimiter
change is caught by nothing else -- both `fidelity_regression` and
`tictactoe_language_coverage` stay green for it.

Declared behavior changes: none -- output-equal move.
