+++ crates/codestory-indexer/src/language_configs.rs
@@ -2,8 +2,8 @@ use super::{
     BASH_GRAPH_QUERY, C_GRAPH_QUERY, CPP_GRAPH_QUERY, CSHARP_GRAPH_QUERY, DART_GRAPH_QUERY,
     GO_GRAPH_QUERY, JAVA_GRAPH_QUERY, JAVASCRIPT_GRAPH_QUERY, LanguageConfig, LanguageRuleset,
     PHP_GRAPH_QUERY, PYTHON_GRAPH_QUERY, RUBY_GRAPH_QUERY, RUST_GRAPH_QUERY, RUST_TAGS_QUERY,
-    SWIFT_GRAPH_QUERY, TSX_GRAPH_QUERY, TSX_TAGS_QUERY, TYPESCRIPT_GRAPH_QUERY,
-    TYPESCRIPT_TAGS_QUERY, languages, make_language_config,
+    TSX_GRAPH_QUERY, TSX_TAGS_QUERY, TYPESCRIPT_GRAPH_QUERY, TYPESCRIPT_TAGS_QUERY, languages,
+    make_language_config,
 };
 use codestory_contracts::language_support::{
     LanguageSupportMode, language_support_profile_for_ext, normalize_extension,
@@ -41,7 +41,6 @@ pub(super) fn get_language_for_ext(ext: &str) -> Option<LanguageConfig> {
         ("ruby", _) => Some(ruby()),
         ("php", _) => Some(php()),
         ("csharp", _) => Some(csharp()),
-        ("swift", _) => Some(swift()),
         ("dart", _) => Some(dart()),
         ("bash", _) => Some(bash()),
         _ => None,
@@ -165,16 +164,6 @@ fn csharp() -> LanguageConfig {
         CSHARP_GRAPH_QUERY,
         None,
         LanguageRuleset::CSharp,
-    )
-}
-
-fn swift() -> LanguageConfig {
-    make_language_config(
-        tree_sitter_swift::LANGUAGE.into(),
-        "swift",
-        SWIFT_GRAPH_QUERY,
-        None,
-        LanguageRuleset::Swift,
     )
 }
