+++ crates/codestory-indexer/src/semantic/mod.rs
@@ -410,9 +410,6 @@ fn dedicated_semantic_resolver(language: &str) -> Option<SemanticResolverKind> {
         "ruby" => Some(SemanticResolverKind::Ruby(RubySemanticResolver)),
         "php" => Some(SemanticResolverKind::Php(PhpSemanticResolver)),
         "csharp" => Some(SemanticResolverKind::CSharp(CSharpSemanticResolver)),
-        "swift" => Some(SemanticResolverKind::Generic(GenericSemanticResolver::new(
-            "swift",
-        ))),
         "dart" => Some(SemanticResolverKind::Generic(GenericSemanticResolver::new(
             "dart",
         ))),
@@ -600,7 +597,6 @@ fn language_family_bucket(language: &'static str) -> &'static str {
         "ruby" => "ruby",
         "php" => "php",
         "csharp" => "csharp",
-        "swift" => "swift",
         "dart" => "dart",
         "bash" => "bash",
         language if is_structural_language_name(language) => "structural",
